#43 apartment와 게시글 erd 연결에 따른 buildSearchContition 수정
