+++ src/services/complaintService.ts
@@ -25,16 +25,12 @@ async function createComplaint(
   });
 }
 
-async function getComplaintList(
-  userId: string,
-  role: USER_ROLE,
-  params: PageParamsType
-) {
+async function getComplaintList(apartmentId: string, params: PageParamsType) {
   const searchCondition = await buildSearchCondition(
     params.page,
     params.limit,
     "",
-    { userId, role }
+    { apartmentId }
   );
   const totalCount = await complaintRepository.getCount({
     where: searchCondition.whereCondition,
